requirements.txt
boto3==1.28.55
Pillow==8.3.2        
